fix(routing): a 404 answers 404
Every not-found in the app answered HTTP 200. A mistyped pay link, an
unknown project id, a route that doesn't exist — all "200 OK" with the
404 UI inside. Search engines index soft 404s as real pages, and the pay
page carried a comment claiming its mistyped-username case already failed
honestly. It didn't.

The cause is structural, and Next's docs state it plainly: "the response
headers are sent before streaming, so the status code cannot be updated
once streaming begins." Two boundaries above every page started that
stream before any page body could run its existence check —

  - <Suspense> around {children} in the root layout
  - src/app/loading.tsx

Both are gone. Measured on a production build (`SELF_HOST=1 npm run
build`, standalone server, real env):

  /this-route-does-not-exist    200 -> 404
  /pay/definitelynotauser       200 -> 404

Two pages read the URL with useSearchParams and were relying on the root
boundary to exist at all; they now carry their own. The reset-password
form moves into its own module so the page is just its boundary.

Not fixed here, and worth being explicit about: routes with their own
loading.tsx (projects/[id], causes/[id], profiles/[username]) still
answer 200, for exactly the same reason one level down. Fixing those
means moving each existence check ahead of its own boundary — 97
route-level files, separate work.

What the root boundary was actually buying: it had NO fallback, so it
flushed chrome around a blank middle, and loading.tsx filled that gap
with a projects-grid skeleton on every route lacking its own — the wrong
skeleton on /about, /settings, everywhere. TTFB now equals full render
rather than shell: /about 0.06s, /pay/mao 0.65s, /discover 0.85s, / 0.87s
on that build. Pages with their own nested boundaries are unaffected
(/causes/... still flushes at 0.41s of a 2.9s total).
